backend/requirements.txt
fastapi
uvicorn
openai
pydantic-settings
python-dotenv
httpx
